fix: Allow empty environment variables
For some reason, os.LookupEnv appears to be playing up on my machine and
is refusing to recognise a variable that has been set to empty string.
The check is more of a matter of convenience, so it can be commented out
until a more permanent solution can be found.