fix: Restore rudimentary colour output (#51)
The YAML log output is lightly coloured using Chroma. Its lexer is
rather conservative and only highlights string values if they are
quoted, which does not appear to be a customisable option in the yaml.v2
package.

This also removes the use of bold titles from the default plain output.