Allow `deploy` without prior `stage`
A pattern that is used with stratus may be to `stage` in branch builds,
and `deploy` in default branch builds.

Upon disasters, like accidental stack deletions, retrying the default branch
to re-create the infrastructure, without worrying about making a temporary branch and staging,
would be a fast path to recovery.